xfs: clean up printing the allocsize option in
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-294.el8
commit-author Christoph Hellwig <hch@lst.de>
commit aa58d4455a11a00c7a3dd39984c98d8793c793e6

Remove superflous cast.

(cherry picked from commit aa58d4455a11a00c7a3dd39984c98d8793c793e6)

+++ fs/xfs/xfs_super.c
@@ -452,7 +452,7 @@ xfs_showargs(
 
 	if (mp->m_flags & XFS_MOUNT_DFLT_IOSIZE)
 		seq_printf(m, ",allocsize=%dk",
-				(int)(1 << mp->m_allocsize_log) >> 10);
+			   (1 << mp->m_allocsize_log) >> 10);
 
 	if (mp->m_logbufs > 0)
 		seq_printf(m, ",logbufs=%d", mp->m_logbufs);
